C2SoftAacEnc: Fix setting SBR mode
SBR Mode value in C2 and the value that aac encoder library accepts
are not same.
SBR Mode in c2 is used to signal both SBR mode and ratio.
So when setting AACENC_SBR_MODE in the encoder library, c2's SBR value
needs to be appropriately converted to values that encoder understands.

Bug: 184898642
Test: atest set aac-sbr-mode using AMediaCodec based application
Test: atest CtsMediaV2TestCases

Change-Id: If4565297bd116230731e454c48f86ba8ca431e2b